geospatial/requirements.txt
pandas==2.1.3
scikit-learn==1.3.2
pyarrow==14.0.1
Flask==3.0.0
Faker==20.1.0
dbldatagen==0.3.5
jmespath==1.0.1
apache-sedona==1.5.1
geopandas==0.14.3
seaborn==0.13.2
